SUP: pass idCpu up on all platforms (save OS/2), handle it in ring-0 on solaris, linux and freebsd.
svn:sync-xref-src-repo-rev: r38863

+++ src/VBox/HostDrivers/Support/win/SUPLib-win.cpp
@@ -1,4 +1,4 @@
-/* $Id: SUPLib-win.cpp 13858 2008-11-05 13:45:41Z noreply@oracle.com $ */
+/* $Id: SUPLib-win.cpp 13865 2008-11-05 14:14:11Z knut.osmundsen@oracle.com $ */
 /** @file
  * VirtualBox Support Library - Windows NT specific parts.
  */
@@ -503,7 +503,7 @@ int suplibOsIOCtl(PSUPLIBDATA pThis, uintptr_t uFunction, void *pvReq, size_t cb
 }
 
 
-int suplibOsIOCtlFast(PSUPLIBDATA pThis, uintptr_t uFunction, unsigned idCpu)
+int suplibOsIOCtlFast(PSUPLIBDATA pThis, uintptr_t uFunction, uintptr_t idCpu)
 {
     /*
      * Issue device I/O control.
